fix: form version should be integer

+++ tasklist/importer-870/src/main/java/io/camunda/tasklist/zeebeimport/v870/processors/es/FormZeebeRecordProcessorElasticSearch.java
@@ -51,7 +51,7 @@ public class FormZeebeRecordProcessorElasticSearch {
       persistForm(
           recordValue.getFormKey(),
           bytesToXml(recordValue.getResource()),
-          (long) recordValue.getVersion(),
+          recordValue.getVersion(),
           recordValue.getTenantId(),
           recordValue.getFormId(),
           false,
@@ -60,7 +60,7 @@ public class FormZeebeRecordProcessorElasticSearch {
       persistForm(
           recordValue.getFormKey(),
           bytesToXml(recordValue.getResource()),
-          (long) recordValue.getVersion(),
+          recordValue.getVersion(),
           recordValue.getTenantId(),
           recordValue.getFormId(),
           true,
@@ -73,7 +73,7 @@ public class FormZeebeRecordProcessorElasticSearch {
   private void persistForm(
       final Long formKey,
       final String schema,
-      final Long version,
+      final int version,
       final String tenantId,
       final String formId,
       final boolean isDelete,
